Create AddPostModal & Shift SVG constants
